refactor(coding-agent): split swarm-coordinator, add WS security, use logger
- Extract SwarmDecisionLoop and SwarmIdleWatchdog into focused modules
  (~600 LOC each) from monolithic swarm-coordinator.ts
- Add getCoordinator(runtime) helper replacing runtime.__swarmCoordinator
- Replace console.log/error with @elizaos/core logger throughout
- Add WS PTY session ownership checks (pty-input/pty-resize require
  pty-subscribe first)
- Add pty-resize dimension validation (finite integers 1-500)
- Add audit logging for writeRaw calls
- Remove wildcard CORS from SSE endpoint (server middleware handles it)
